feat(platform): deliver multi-module runtime and UX upgrades
This commit consolidates broad platform work across the web app, shared libraries, and Tauri backend while keeping static checks clean.

Highlights:

- External agent and routing: expanded external-agent management flow, added config normalization, and wired provider initialization paths.

- TTS pipeline: introduced shared API utilities, new Cartesia/Deepgram providers, provider status routes, and orchestration adapters.

- Workflow/webhook handling: added unified webhook entrypoints and improved route handling for workflow triggers.

- Speedpass and learning: added mode/settings dialog flows, tutorial listing support, and learning adapters (QTI/xAPI/textbook extraction).

- Editor workbench and command routing: added editor workbench/runtime helpers, context-driven command integration, and associated type definitions.

- Theme/background hardening: added theme resolver, migration/safety/background activity helpers with tests.

- Persistence/storage: added persistence primitives and backup/crypto/feature-flag persistence utilities.

- Plugin/devtools/runtime integrations: added dev server/hot-reload clients and expanded plugin/runtime surface area.

- Tauri backend: added chat runtime modules plus compatibility, LSP/devtools, scheduler metadata, and timeline rendering commands.

- Test coverage updates: refreshed/added tests across chat, scheduler, speedpass, themes, storage, and editor-related modules.

Quality and fixes:

- Resolved React hook dependency warnings in artifact panel, canvas panel, and editor/command-palette memoization paths.

- Verification executed successfully: pnpm lint; pnpm exec tsc --noEmit; targeted jest suites for artifact panel, command palette, and monaco sandpack editor.

+++ src-tauri/src/screenshot/annotator.rs
@@ -400,11 +400,11 @@ impl ScreenshotAnnotator {
             for py in (cy_i - ry_i).max(0)..(cy_i + ry_i + 1).min(self.image_height as i32) {
                 let dy = (py - cy_i) as f64;
                 // x^2/rx^2 + y^2/ry^2 = 1 => x = rx * sqrt(1 - y^2/ry^2)
-                let ratio = 1.0 - (dy * dy) / (ry as f64 * ry as f64);
+                let ratio = 1.0 - (dy * dy) / (ry * ry);
                 if ratio < 0.0 {
                     continue;
                 }
-                let half_width = (rx as f64 * ratio.sqrt()) as i32;
+                let half_width = (rx * ratio.sqrt()) as i32;
                 for px in
                     (cx_i - half_width).max(0)..(cx_i + half_width + 1).min(self.image_width as i32)
                 {
@@ -423,11 +423,10 @@ impl ScreenshotAnnotator {
                     let dx = (px - cx_i) as f64;
                     let dy = (py - cy_i) as f64;
                     // Normalized distance from ellipse boundary
-                    let dist =
-                        (dx * dx) / (rx as f64 * rx as f64) + (dy * dy) / (ry as f64 * ry as f64);
+                    let dist = (dx * dx) / (rx * rx) + (dy * dy) / (ry * ry);
                     let dist_sqrt = dist.sqrt();
                     // Check if pixel is on the ellipse border (within stroke width)
-                    let half_stroke = stroke as f64 / (2.0 * rx.min(ry) as f64);
+                    let half_stroke = stroke as f64 / (2.0 * rx.min(ry));
                     if (dist_sqrt - 1.0).abs() <= half_stroke {
                         self.set_pixel(pixels, px as u32, py as u32, r, g, b, a);
                     }
